requirements.txt
robotframework==6.1.1
robotframework-browser==16.0.0
robotframework-requests
fastapi==0.108.0
uvicorn==0.25.0
requests
pydantic_settings
opencv-python
numpy
tensorflow==2.15.0
scikit-learn
mediapipe
